organization: move proxy class to its own source and header files
