+++ src/axom/inlet/Table.hpp
@@ -118,128 +118,7 @@ struct has_FromInlet_specialization<
 
 }  // namespace detail
 
-/*!
- *******************************************************************************
- * \class Proxy
- *
- * \brief Provides a uniform interface for access and conversion to primitive
- * and user-defined types
- *
- * \see Inlet Field
- * \see Inlet Table
- *******************************************************************************
- */
-class Proxy
-{
-public:
-  Proxy() = default;
-  /*!
-   *******************************************************************************
-   * \brief Constructs a proxy view onto a table
-   * 
-   * \param [in] table The table to construct a proxy into
-   *******************************************************************************
-   */
-  Proxy(Table& table) : m_table(&table) { }
-
-  /*!
-   *******************************************************************************
-   * \brief Constructs a proxy view onto a field
-   * 
-   * \param [in] field The field to construct a proxy into
-   *******************************************************************************
-   */
-  Proxy(Field& field) : m_field(&field) { }
-
-  /*!
-   *******************************************************************************
-   * \brief Returns an object from the proxy
-   * 
-   * \tparam T The type of the object to retrieve
-   * \return The retrieved object
-   * \note Implicit conversions are defined only for primitive types - to 
-   * convert to a user-defined type, use an explicit conversion with static_cast
-   * or T{}
-   *******************************************************************************
-   */
-  template <typename T,
-            typename SFINAE =
-              typename std::enable_if<detail::is_inlet_primitive<T>::value ||
-                                      detail::is_inlet_primitive_array<T>::value>::type>
-  operator T()
-  {
-    return get<T>();
-  }
-
-  template <typename T>
-  explicit operator T()
-  {
-    return get<T>();
-  }
-
-  /*!
-   *****************************************************************************
-   * \brief Return whether a subobject with the given name is present in 
-   * the table referred to by the calling proxy.
-   *
-   * \return Boolean value indicating whether this Table's subtree contains a
-   * Field or Table with the given name.
-   *****************************************************************************
-   */
-  bool contains(const std::string& name);
-
-  /*!
-   *****************************************************************************
-   * \brief Returns the type of the stored value
-   * 
-   * \return The type
-   * \see InletType
-   *****************************************************************************
-   */
-  InletType type() const;
-
-  /*!
-   *******************************************************************************
-   * \brief Obtains a proxy view into the proxy for either a Field/Table subobject
-   * 
-   * Returns a reference via a lightweight proxy object to the element in the 
-   * datastore at the index specified by the name.  This can be a field 
-   * or a table.
-   * 
-   * \param [in] name The name of the subobject
-   * \return A view onto the subobject
-   *******************************************************************************
-   */
-  Proxy operator[](const std::string& name);
-
-  /*!
-   *******************************************************************************
-   * \brief Returns a user-defined type from the proxy
-   * 
-   * \tparam T The type of the object to retrieve
-   * \return The retrieved object
-   * \pre The Proxy must refer to a table object
-   *******************************************************************************
-   */
-  template <typename T>
-  typename std::enable_if<!detail::is_inlet_primitive<T>::value, T>::type get();
-
-  /*!
-   *******************************************************************************
-   * \brief Returns a primitive type from the proxy
-   * 
-   * \tparam T The type of the object to retrieve
-   * \return The retrieved object
-   * \pre The Proxy must refer to a field object
-   *******************************************************************************
-   */
-  template <typename T>
-  typename std::enable_if<detail::is_inlet_primitive<T>::value, T>::type get();
-
-private:
-  Table* m_table = nullptr;
-  Field* m_field = nullptr;
-};
+class Proxy;
 
 /*!
  *******************************************************************************
@@ -853,24 +732,6 @@ private:
   std::unordered_map<std::string, std::shared_ptr<Field>> m_fieldChildren;
   std::function<bool()> m_verifier;
 };
-
-template <typename T>
-typename std::enable_if<detail::is_inlet_primitive<T>::value, T>::type Proxy::get()
-{
-  SLIC_ASSERT_MSG(
-    m_field != nullptr,
-    "[Inlet] Tried to read a primitive type from a Proxy containing a table");
-  return m_field->get<T>();
-}
-
-template <typename T>
-typename std::enable_if<!detail::is_inlet_primitive<T>::value, T>::type Proxy::get()
-{
-  SLIC_ASSERT_MSG(m_table != nullptr,
-                  "[Inlet] Tried to read a user-defined type from a Proxy "
-                  "containing a single field");
-  return m_table->get<T>();
-}
 
 }  // end namespace inlet
 }  // end namespace axom
